Moved array-expression specific rewrite code to separate module.  Started working on array expression lowering.
